trying to make files work

+++ src/job-control/put_job_in_foreground.c
@@ -6,7 +6,7 @@
 /*   By: jhalford <jack@crans.org>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/13 14:58:36 by jhalford          #+#    #+#             */
-/*   Updated: 2017/03/03 18:52:03 by jhalford         ###   ########.fr       */
+/*   Updated: 2017/03/03 18:57:58 by jhalford         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,21 +16,28 @@ int		put_job_in_foreground(t_job *j, int cont)
 {
 	t_jobc	*jobc;
 
-	jobc = &data_singleton()->jobc;
-	tcsetpgrp(STDIN, j->pgid);
-	tcsetattr(STDIN, TCSADRAIN, &jobc->shell_tmodes);
-	if (cont)
+	if (SH_HAS_JOBC(data_singleton()->opts))
 	{
-		tcsetattr(STDIN, TCSADRAIN, &j->tmodes);
-		if (kill(-j->pgid, SIGCONT) < 0)
-			DG("kill(SIGCONT) failed");
+		jobc = &data_singleton()->jobc;
+
+		tcsetpgrp(STDIN, j->pgid);
+		tcsetattr(STDIN, TCSADRAIN, &jobc->shell_tmodes);
+		if (cont)
+		{
+			tcsetattr(STDIN, TCSADRAIN, &j->tmodes);
+			if (kill(-j->pgid, SIGCONT) < 0)
+				DG("kill(SIGCONT) failed");
+		}
+		job_wait(j->id);
+		job_remove(j->id);
+
+		tcsetpgrp(STDIN, jobc->shell_pgid);
+
+		tcgetattr(STDIN, &j->tmodes);
+		tcsetattr(STDIN, TCSADRAIN, &jobc->shell_tmodes);
+		tcsetattr(STDIN, TCSADRAIN, &jobc->shell_tmodes);
 	}
-	job_wait(j->id);
-	job_remove(j->id);
-
-	tcsetpgrp(STDIN, jobc->shell_pgid);
-
-	tcgetattr(STDIN, &j->tmodes);
-	tcsetattr(STDIN, TCSADRAIN, &jobc->shell_tmodes);
+	else
+		job_wait(j->id);
 	return (0);
 }
